refactor: factor-out functions from generate_hashes.py to own utils.py
as to aid new script which will be used in ci that does it all
